+++ routes/apis.js
@@ -4,7 +4,7 @@ const passport = require('../config/passport')
 const restController = require('../controllers/api/restController')
 const userController = require('../controllers/api/userController')
 const businessController = require('../controllers/api/businessController')
-const businessService = require('../services/businessService')
+const orderController = require('../controllers/api/orderController')
 const multer = require('multer')
 const upload = multer({ dest: 'temp/' })
 
@@ -27,19 +27,21 @@ router.post('/signin', userController.signIn)
 router.get('/get_current_user', authenticated, userController.getUser)
 router.get('/home', restController.getRestaurants)
 router.get('/restaurant/:id', restController.getRestaurant)
-router.get('/reservation/:id', restController.getMeals)
+router.get('/restaurant/:id/meals', restController.getMeals)
 
 //common_user
 router.post('/comment', authenticated, userController.postComment)
-router.post('/reservation/:id', authenticated, userController.postReservation)
 router.get('/member/:id/orders', authenticated, userController.getOrders)
 router.put('/member/edit', authenticated, upload.single('image'), userController.putUser)
 router.get('/member/info', authenticated, userController.getUserInfo)
 
+//orders
+router.post('/order/:id', authenticated, orderController.postOrder)
+
 //business_user
 router.get('/business/restaurant', authenticated, authenticatedBusiness, businessController.getRestaurant)
 router.get('/business/menu', authenticated, authenticatedBusiness, businessController.getMenu)
-router.put('/business/restaurant', authenticated, authenticatedBusiness, upload.single('image'), businessService.putRestaurant)
+router.put('/business/restaurant', authenticated, authenticatedBusiness, upload.single('image'), businessController.putRestaurant)
 router.put('/business/menu', authenticated, authenticatedBusiness, upload.single('image'), businessController.putMenu)
 router.post('/business/meal', authenticated, authenticatedBusiness, upload.single('image'), businessController.postMeal)
