+++ services/userService.js
@@ -13,7 +13,7 @@ const moment = require('moment')
 const imgur = require('imgur-node-api')
 const IMGUR_CLIENT_ID = process.env.IMGUR_CLIENT_ID
 
-let userService = {
+const userService = {
   getUser: (req, res, callback) => {
     return User.findByPk(req.user.dataValues.id, {
       include: [
@@ -53,52 +53,6 @@ let userService = {
         })
       })
 
-    })
-      .catch(err => res.send(err))
-  },
-  postReservation: (req, res, callback) => {
-    if (!req.body.info.seat || !req.body.info.time || !req.body.info.name || !req.body.info.phone || !req.body.info.date || !req.body.info.totalPrice) {
-      return callback({ status: 'error', message: '所有欄位為必填' })
-    }
-    const { time, name, phone, date, note, totalPrice } = req.body.info
-    const seatCount = req.body.info.seat
-
-
-    RestaurantSeat.findOne({
-      where: { RestaurantId: req.params.id }
-    }).then((seat) => {
-      Order.create({
-        UserId: Number(req.user.dataValues.id),
-        RestaurantSeatId: Number(seat.dataValues.id),
-        time: time.toString(),
-        peopleCount: seatCount,
-        note: note,
-        reserve_name: name,
-        reserve_phone: phone,
-        date: date.toString(),
-        totalPrice: Number(totalPrice)
-      }).then((order) => {
-        let meals = req.body.orders
-        OrderItem.bulkCreate(
-          Array.from({ length: meals.length }).map((_, index) => ({
-            MealId: Number(meals[index].id),
-            OrderId: Number(order.dataValues.id),
-            quantity: Number(meals[index].quantity)
-          }))
-        ).then((bulk) => {
-          let restSeat = seat.seat - order.peopleCount
-          seat.update({
-            seat: restSeat
-          }).then(() => {
-            return callback({
-              status: 'success',
-              message: '訂位&訂餐成功',
-              order: order
-            })
-          })
-        })
-
-      })
     })
       .catch(err => res.send(err))
   },
